Moved AABB4 testing
